+++ internal/cluster/clusterpb/JoinReply.go
@@ -89,8 +89,28 @@ func (rcv *JoinReply) LeaderAddr() []byte {
 	return nil
 }
 
+func (rcv *JoinReply) PeerSpkis(obj *SPKIBytes, j int) bool {
+	o := flatbuffers.UOffsetT(rcv._tab.Offset(14))
+	if o != 0 {
+		x := rcv._tab.Vector(o)
+		x += flatbuffers.UOffsetT(j) * 4
+		x = rcv._tab.Indirect(x)
+		obj.Init(rcv._tab.Bytes, x)
+		return true
+	}
+	return false
+}
+
+func (rcv *JoinReply) PeerSpkisLength() int {
+	o := flatbuffers.UOffsetT(rcv._tab.Offset(14))
+	if o != 0 {
+		return rcv._tab.VectorLen(o)
+	}
+	return 0
+}
+
 func JoinReplyStart(builder *flatbuffers.Builder) {
-	builder.StartObject(5)
+	builder.StartObject(6)
 }
 func JoinReplyAddAccepted(builder *flatbuffers.Builder, accepted bool) {
 	builder.PrependBoolSlot(0, accepted, false)
@@ -106,6 +126,12 @@ func JoinReplyAddLeaderId(builder *flatbuffers.Builder, leaderId flatbuffers.UOf
 }
 func JoinReplyAddLeaderAddr(builder *flatbuffers.Builder, leaderAddr flatbuffers.UOffsetT) {
 	builder.PrependUOffsetTSlot(4, flatbuffers.UOffsetT(leaderAddr), 0)
+}
+func JoinReplyAddPeerSpkis(builder *flatbuffers.Builder, peerSpkis flatbuffers.UOffsetT) {
+	builder.PrependUOffsetTSlot(5, flatbuffers.UOffsetT(peerSpkis), 0)
+}
+func JoinReplyStartPeerSpkisVector(builder *flatbuffers.Builder, numElems int) flatbuffers.UOffsetT {
+	return builder.StartVector(4, numElems, 4)
 }
 func JoinReplyEnd(builder *flatbuffers.Builder) flatbuffers.UOffsetT {
 	return builder.EndObject()